implement generatePopulation method in PopulationGenerator

+++ src/main/java/com/example/epidemicsimulator/model/Population.java
@@ -4,8 +4,8 @@ import lombok.Builder;
 
 @Builder
 public record Population(
-        int numberOfInfectedPeople,
-        int numberOfHealthyPeople,
-        int numberOfDiedPeople,
-        int numberOfRecoveredPeople) {
+        long numberOfInfectedPeople,
+        long numberOfHealthyPeople,
+        long numberOfDiedPeople,
+        long numberOfRecoveredPeople) {
 }
